Minor Changes
The changes made in these files are just for better readability. It has been ensured that no line has more than 80 characters. Also, the header files have been added according to whether they are user-defined or are a part of C++ package. There is no significant change to the code.